fix: address new storage review comments
