+++ cache/doc.go
@@ -1,3 +1,3 @@
 // Package cache provides a backend-neutral cache abstraction with explicit
-// factory registration and an in-memory default implementation.
+// factory registration and an in-memory backend for the default provider name.
 package cache
