+++ database/component_test.go
@@ -4,13 +4,15 @@ import (
 	"context"
 	"testing"
 
-	"gorm.io/driver/sqlite"
+	"gorm.io/gorm"
 
 	"github.com/kbukum/gokit/component"
 	"github.com/kbukum/gokit/logger"
 )
 
-// TestComponent_Name tests that the component returns the correct name
+func fakeDriver(string) gorm.Dialector { return nil }
+
+// TestComponent_Name tests that the component returns the correct name.
 func TestComponent_Name(t *testing.T) {
 	cfg := Config{
 		Enabled: true,
@@ -25,7 +27,7 @@ func TestComponent_Name(t *testing.T) {
 	}
 }
 
-// TestComponent_Interface tests that Component satisfies component.Component
+// TestComponent_Interface tests that Component satisfies component.Component.
 func TestComponent_Interface(t *testing.T) {
 	cfg := Config{
 		Enabled: true,
@@ -37,40 +39,7 @@ func TestComponent_Interface(t *testing.T) {
 	var _ component.Component = comp
 }
 
-// TestComponent_Lifecycle tests basic start/stop lifecycle
-func TestComponent_Lifecycle(t *testing.T) {
-	cfg := Config{
-		Enabled: true,
-		DSN:     ":memory:",
-	}
-	cfg.ApplyDefaults()
-	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
-
-	ctx := context.Background()
-
-	// Component should not have db before start
-	if db := comp.DB(); db != nil {
-		t.Error("DB() should be nil before Start")
-	}
-
-	// Start should succeed
-	if err := comp.Start(ctx); err != nil {
-		t.Fatalf("Start() failed: %v", err)
-	}
-
-	// Component should have db after start
-	if db := comp.DB(); db == nil {
-		t.Error("DB() should not be nil after Start")
-	}
-
-	// Stop should succeed
-	if err := comp.Stop(ctx); err != nil {
-		t.Fatalf("Stop() failed: %v", err)
-	}
-}
-
-// TestComponent_WithDriver tests custom driver function
+// TestComponent_WithDriver tests custom driver function wiring without importing a driver SDK.
 func TestComponent_WithDriver(t *testing.T) {
 	cfg := Config{
 		Enabled: true,
@@ -80,29 +49,9 @@ func TestComponent_WithDriver(t *testing.T) {
 	log := logger.NewDefault("test")
 	comp := NewComponent(cfg, log)
 
-	// Custom driver that wraps sqlite
-	customDriver := sqlite.Open
-
-	// WithDriver should return the component for chaining
-	result := comp.WithDriver(customDriver)
+	result := comp.WithDriver(fakeDriver)
 	if result != comp {
 		t.Error("WithDriver() should return the component for method chaining")
-	}
-
-	ctx := context.Background()
-
-	// Should start successfully with custom driver
-	if err := comp.Start(ctx); err != nil {
-		t.Fatalf("Start() with custom driver failed: %v", err)
-	}
-
-	// Component should have db
-	if db := comp.DB(); db == nil {
-		t.Error("DB() should not be nil after Start with custom driver")
-	}
-
-	if err := comp.Stop(ctx); err != nil {
-		t.Fatalf("Stop() failed: %v", err)
 	}
 }
 
@@ -130,110 +79,35 @@ func TestDriverRegistryNoSideEffects(t *testing.T) {
 	}
 }
 
-// TestComponent_WithAutoMigrate_Enabled tests auto-migration when enabled
-func TestComponent_WithAutoMigrate_Enabled(t *testing.T) {
-	cfg := Config{
-		Enabled:     true,
-		DSN:         ":memory:",
-		AutoMigrate: true,
-	}
-	cfg.ApplyDefaults()
-	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
-
-	// Simple test model
-	type User struct {
-		ID   uint
-		Name string
-	}
-
-	// Register model for auto-migration
-	comp.WithAutoMigrate(&User{})
-
-	ctx := context.Background()
-
-	// Start should succeed and run auto-migration
-	if err := comp.Start(ctx); err != nil {
-		t.Fatalf("Start() failed: %v", err)
-	}
-
-	// Verify table was created by checking if we can query it
-	if !comp.DB().GormDB.Migrator().HasTable(&User{}) {
-		t.Error("User table should have been migrated")
-	}
-
-	if err := comp.Stop(ctx); err != nil {
-		t.Fatalf("Stop() failed: %v", err)
-	}
-}
-
-// TestComponent_WithAutoMigrate_Disabled tests auto-migration when disabled
-func TestComponent_WithAutoMigrate_Disabled(t *testing.T) {
-	cfg := Config{
-		Enabled:     true,
-		DSN:         ":memory:",
-		AutoMigrate: false,
-	}
-	cfg.ApplyDefaults()
-	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
-
-	type User struct {
-		ID   uint
-		Name string
-	}
-
-	// Register model but auto-migrate is disabled
-	comp.WithAutoMigrate(&User{})
-
-	ctx := context.Background()
-
-	// Start should succeed but table should not be created
-	if err := comp.Start(ctx); err != nil {
-		t.Fatalf("Start() failed: %v", err)
-	}
-
-	// Table should not exist
-	if comp.DB().GormDB.Migrator().HasTable(&User{}) {
-		t.Error("User table should not have been migrated when AutoMigrate is false")
-	}
-
-	if err := comp.Stop(ctx); err != nil {
-		t.Fatalf("Stop() failed: %v", err)
-	}
-}
-
-// TestComponent_WithAutoMigrate_Chaining tests that WithAutoMigrate returns component
+// TestComponent_WithAutoMigrate_Chaining tests that WithAutoMigrate returns component.
 func TestComponent_WithAutoMigrate_Chaining(t *testing.T) {
 	cfg := Config{
 		Enabled: true,
 		DSN:     ":memory:",
 	}
 	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
+	comp := NewComponent(cfg, log).WithDriver(fakeDriver)
 
 	type User struct {
 		ID uint
 	}
 
-	// WithAutoMigrate should return the component for chaining
 	result := comp.WithAutoMigrate(&User{})
 	if result != comp {
 		t.Error("WithAutoMigrate() should return the component for method chaining")
 	}
 }
 
-// TestComponent_Health_BeforeStart tests health check before component starts
+// TestComponent_Health_BeforeStart tests health check before component starts.
 func TestComponent_Health_BeforeStart(t *testing.T) {
 	cfg := Config{
 		Enabled: true,
 		DSN:     ":memory:",
 	}
 	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
+	comp := NewComponent(cfg, log).WithDriver(fakeDriver)
 
 	ctx := context.Background()
-
 	health := comp.Health(ctx)
 
 	if health.Name != "database" {
@@ -247,38 +121,7 @@ func TestComponent_Health_BeforeStart(t *testing.T) {
 	}
 }
 
-// TestComponent_Health_AfterStart tests health check after component starts
-func TestComponent_Health_AfterStart(t *testing.T) {
-	cfg := Config{
-		Enabled: true,
-		DSN:     ":memory:",
-	}
-	cfg.ApplyDefaults()
-	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
-
-	ctx := context.Background()
-
-	// Start the component
-	if err := comp.Start(ctx); err != nil {
-		t.Fatalf("Start() failed: %v", err)
-	}
-
-	health := comp.Health(ctx)
-
-	if health.Name != "database" {
-		t.Errorf("Health Name = %q, want %q", health.Name, "database")
-	}
-	if health.Status != component.StatusHealthy {
-		t.Errorf("Health Status = %q, want %q", health.Status, component.StatusHealthy)
-	}
-
-	if err := comp.Stop(ctx); err != nil {
-		t.Fatalf("Stop() failed: %v", err)
-	}
-}
-
-// TestComponent_Describe tests the Describe method
+// TestComponent_Describe tests the Describe method.
 func TestComponent_Describe(t *testing.T) {
 	cfg := Config{
 		Enabled:      true,
@@ -300,43 +143,12 @@ func TestComponent_Describe(t *testing.T) {
 	if desc.Details == "" {
 		t.Error("Describe Details should not be empty")
 	}
-	// Verify DSN is masked
 	if desc.Details != "" && desc.Details[0:3] != "DSN" {
 		t.Error("Describe Details should start with DSN")
 	}
 }
 
-// TestNewWithDialector tests NewWithDialector function
-func TestNewWithContext(t *testing.T) {
-	cfg := Config{
-		Enabled: true,
-		DSN:     ":memory:",
-	}
-	cfg.ApplyDefaults()
-	log := logger.NewDefault("test")
-
-	dialector := sqlite.Open(cfg.DSN)
-
-	db, err := NewWithContext(context.Background(), dialector, cfg, log)
-	if err != nil {
-		t.Fatalf("NewWithContext() failed: %v", err)
-	}
-
-	if db == nil {
-		t.Error("NewWithContext() returned nil DB")
-	}
-
-	// Verify database is functional
-	if err := db.Ping(); err != nil {
-		t.Errorf("Ping() failed: %v", err)
-	}
-
-	if err := db.Close(); err != nil {
-		t.Fatalf("Close() failed: %v", err)
-	}
-}
-
-// TestNewWithContext_InvalidType tests NewWithContext with invalid dialector
+// TestNewWithContext_InvalidType tests NewWithContext with an invalid dialector type.
 func TestNewWithContext_InvalidType(t *testing.T) {
 	cfg := Config{
 		Enabled: true,
@@ -345,42 +157,37 @@ func TestNewWithContext_InvalidType(t *testing.T) {
 	cfg.ApplyDefaults()
 	log := logger.NewDefault("test")
 
-	// Pass an invalid dialector type (string instead of gorm.Dialector)
 	invalidDialector := "not-a-dialector"
-
 	db, err := NewWithContext(context.Background(), invalidDialector, cfg, log)
 
 	if err == nil {
 		t.Error("NewWithContext() should return an error for invalid dialector type")
 	}
-
 	if db != nil {
 		t.Error("NewWithContext() should return nil DB on error")
 	}
-
 	if errMsg := err.Error(); errMsg == "" {
 		t.Error("Error message should not be empty")
 	}
 }
 
-// TestComponent_Stop_BeforeStart tests Stop before Start is called
+// TestComponent_Stop_BeforeStart tests Stop before Start is called.
 func TestComponent_Stop_BeforeStart(t *testing.T) {
 	cfg := Config{
 		Enabled: true,
 		DSN:     ":memory:",
 	}
 	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
+	comp := NewComponent(cfg, log).WithDriver(fakeDriver)
 
 	ctx := context.Background()
 
-	// Stop should be idempotent and not error when db is nil
 	if err := comp.Stop(ctx); err != nil {
 		t.Fatalf("Stop() before Start() should not error: %v", err)
 	}
 }
 
-// TestComponent_ChainedMethods tests that methods can be chained
+// TestComponent_ChainedMethods tests that methods can be chained.
 func TestComponent_ChainedMethods(t *testing.T) {
 	cfg := Config{
 		Enabled:     true,
@@ -394,46 +201,16 @@ func TestComponent_ChainedMethods(t *testing.T) {
 		ID uint
 	}
 
-	// Chain methods
 	comp := NewComponent(cfg, log).
-		WithDriver(sqlite.Open).
+		WithDriver(fakeDriver).
 		WithAutoMigrate(&User{})
 
 	if comp == nil {
 		t.Error("Chained methods should return component")
 	}
-
-	ctx := context.Background()
-	if err := comp.Start(ctx); err != nil {
-		t.Fatalf("Start() failed: %v", err)
-	}
-
-	if err := comp.Stop(ctx); err != nil {
-		t.Fatalf("Stop() failed: %v", err)
-	}
 }
 
-// TestComponent_StartWithInvalidDSN tests Start with invalid DSN
-func TestComponent_StartWithInvalidDSN(t *testing.T) {
-	cfg := Config{
-		Enabled: true,
-		DSN:     "/invalid/path/to/db.db",
-	}
-	cfg.ApplyDefaults()
-	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
-
-	ctx := context.Background()
-
-	// Start might fail due to invalid path
-	err := comp.Start(ctx)
-	// Either succeeds or fails gracefully - just verify it's handled
-	if err != nil && err.Error() == "" {
-		t.Error("Error message should not be empty")
-	}
-}
-
-// TestComponent_DB_ReturnsNilBeforeStart tests DB() returns nil before Start
+// TestComponent_DB_ReturnsNilBeforeStart tests DB() returns nil before Start.
 func TestComponent_DB_ReturnsNilBeforeStart(t *testing.T) {
 	cfg := Config{
 		Enabled: true,
@@ -447,32 +224,7 @@ func TestComponent_DB_ReturnsNilBeforeStart(t *testing.T) {
 	}
 }
 
-// TestComponent_DB_ReturnsValueAfterStart tests DB() returns value after Start
-func TestComponent_DB_ReturnsValueAfterStart(t *testing.T) {
-	cfg := Config{
-		Enabled: true,
-		DSN:     ":memory:",
-	}
-	cfg.ApplyDefaults()
-	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
-
-	ctx := context.Background()
-	if err := comp.Start(ctx); err != nil {
-		t.Fatalf("Start() failed: %v", err)
-	}
-
-	db := comp.DB()
-	if db == nil {
-		t.Error("DB() should not return nil after Start")
-	}
-
-	if err := comp.Stop(ctx); err != nil {
-		t.Fatalf("Stop() failed: %v", err)
-	}
-}
-
-// TestComponent_Disabled tests component behavior when Enabled=false
+// TestComponent_Disabled tests component behavior when Enabled=false.
 func TestComponent_Disabled(t *testing.T) {
 	cfg := Config{
 		Enabled: false,
@@ -480,21 +232,17 @@ func TestComponent_Disabled(t *testing.T) {
 	}
 	cfg.ApplyDefaults()
 	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
+	comp := NewComponent(cfg, log).WithDriver(fakeDriver)
 
 	ctx := context.Background()
 
-	// Start should succeed but skip initialization
 	if err := comp.Start(ctx); err != nil {
 		t.Fatalf("Start() with Enabled=false should not error: %v", err)
 	}
-
-	// DB should remain nil
 	if db := comp.DB(); db != nil {
 		t.Error("DB() should be nil when component is disabled")
 	}
 
-	// Health should return healthy with "disabled" message
 	health := comp.Health(ctx)
 	if health.Status != component.StatusHealthy {
 		t.Errorf("Health Status = %q, want %q", health.Status, component.StatusHealthy)
@@ -502,59 +250,26 @@ func TestComponent_Disabled(t *testing.T) {
 	if health.Message != "disabled" {
 		t.Errorf("Health Message = %q, want %q", health.Message, "disabled")
 	}
-
-	// Stop should be safe
 	if err := comp.Stop(ctx); err != nil {
 		t.Fatalf("Stop() with Enabled=false should not error: %v", err)
 	}
 }
 
-// TestComponent_EnabledDefaultTrue tests that Enabled defaults to false (zero value)
+// TestComponent_EnabledDefaultBehavior tests that Enabled defaults to false.
 func TestComponent_EnabledDefaultBehavior(t *testing.T) {
 	cfg := Config{
-		// Enabled not set - defaults to false
 		DSN: ":memory:",
 	}
 	cfg.ApplyDefaults()
 	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
+	comp := NewComponent(cfg, log).WithDriver(fakeDriver)
 
 	ctx := context.Background()
 
-	// Should skip initialization when Enabled is false (default)
 	if err := comp.Start(ctx); err != nil {
 		t.Fatalf("Start() should not error with default Enabled=false: %v", err)
 	}
-
 	if db := comp.DB(); db != nil {
 		t.Error("DB() should be nil when Enabled defaults to false")
-	}
-}
-
-// TestComponent_ContextInHealthCheck tests context handling in Health()
-func TestComponent_ContextInHealthCheck(t *testing.T) {
-	cfg := Config{
-		Enabled: true,
-		DSN:     ":memory:",
-	}
-	cfg.ApplyDefaults()
-	log := logger.NewDefault("test")
-	comp := NewComponent(cfg, log).WithDriver(sqlite.Open)
-
-	ctx := context.Background()
-	if err := comp.Start(ctx); err != nil {
-		t.Fatalf("Start() failed: %v", err)
-	}
-	defer comp.Stop(ctx)
-
-	// Create a canceled context
-	canceledCtx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	// Health check with canceled context should still work but might fail
-	health := comp.Health(canceledCtx)
-	// The health check should at least not panic
-	if health.Name != "database" {
-		t.Errorf("Health Name = %q, want %q", health.Name, "database")
 	}
 }
